requirements.txt
# Hospital Patient Records Management System - Python Dependencies

# Web Framework
flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.3
flask-migrate==4.0.5
flask-restful==0.3.10

# Database
pymongo==4.5.0
mongoengine==0.27.0

# Authentication & Security
bcrypt==4.0.1
pyjwt==2.8.0
cryptography==41.0.4

# Data Validation & Serialization
marshmallow==3.20.1
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0

# Email Services
flask-mail==0.9.1

# Environment Configuration
python-dotenv==1.0.0

# Date & Time Utilities
python-dateutil==2.8.2
pytz==2023.3

# File Handling
pillow==10.0.1

# Testing
pytest==7.4.2
pytest-flask==1.2.0

# Development Tools
black==23.7.0
flake8==6.0.0

# Monitoring & Logging
gunicorn==21.2.0
sentry-sdk==1.30.0
